tools/mkbinheader: Support kernel binary header
Support kernel binary header for binary update.
 - CRC32 : uint32_t, 4bytes
 - Header size : uint16_t, 2bytes
 - Binary version(YYMMDD) : uint32_t, 4bytes
 - Binary size : uint32_t, 4bytes
 - Secure header size : uint16_t, 2bytes
 - Secure header : #bytes.

+++ os/kernel/binary_manager/binary_manager.h
@@ -148,7 +148,7 @@ struct binmgr_kinfo_s {
 	uint32_t part_count;
 	uint32_t part_size[KERNEL_BIN_COUNT];
 	int8_t part_num[KERNEL_BIN_COUNT];
-	float version;
+	uint32_t version;
 };
 typedef struct binmgr_kinfo_s binmgr_kinfo_t;
